backend/requirements.txt
psycopg2-binary==2.9.9     # PostgreSQL
pymongo==4.12.1            # MongoDB
neo4j==5.28.1              # Neo4j Python driver
dnspython==2.7.0           # Required by pymongo
faker==25.2.0              # For data generation (optional but useful)
